move rendering logic from engine to renderer
