+++ CMakeLists.txt
@@ -127,15 +127,16 @@ set (SourceVQE
     "Source/Engine/AssetLoader.h"
     "Source/Engine/GPUMarker.h"
     "Source/Engine/EnvironmentMap.h"
+    "Source/Engine/LoadingScreen.h"
 
     "Source/Engine/Main.cpp"
     "Source/Engine/VQEngine_Main.cpp"
     "Source/Engine/VQEngine_Render.cpp"
-    "Source/Engine/VQEngine_RenderCommands.cpp"
     "Source/Engine/VQEngine_Update.cpp"
     "Source/Engine/VQEngine_Input.cpp"
     "Source/Engine/VQEngine_Simulation.cpp"
     "Source/Engine/EnvironmentMap.cpp"
+    "Source/Engine/LoadingScreen.cpp"
     "Source/Engine/Math.cpp"
     "Source/Engine/Culling.cpp"
     "Source/Engine/AssetLoader.cpp"
