+++ Source/Engine/LoadingScreen.cpp
@@ -1,0 +1,31 @@
+//	VQE
+//	Copyright(C) 2025  - Volkan Ilbeyli
+//
+//	This program is free software : you can redistribute it and / or modify
+//	it under the terms of the GNU General Public License as published by
+//	the Free Software Foundation, either version 3 of the License, or
+//	(at your option) any later version.
+//
+//	This program is distributed in the hope that it will be useful,
+//	but WITHOUT ANY WARRANTY; without even the implied warranty of
+//	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+//	GNU General Public License for more details.
+//
+//	You should have received a copy of the GNU General Public License
+//	along with this program.If not, see <http://www.gnu.org/licenses/>.
+//
+//	Contact: volkanilbeyli@gmail.com
+
+#include "LoadingScreen.h"
+#include "Libs/VQUtils/Source/utils.h"
+#include <cassert>
+
+SRV_ID FLoadingScreenData::GetSelectedLoadingScreenSRV_ID() const
+{
+	assert(SelectedLoadingScreenSRVIndex < SRVs.size());
+	return SRVs[SelectedLoadingScreenSRVIndex];
+}
+void FLoadingScreenData::RotateLoadingScreenImageIndex()
+{
+	SelectedLoadingScreenSRVIndex = (int)MathUtil::RandU(0, (int)SRVs.size());
+}
